rename package from chocofamily to example
