--- app/src/main/java/kz/chocofamily/coroutinelesson/presentation/MainActivity.kt
@@ -1,13 +1,0 @@
-package kz.chocofamily.coroutinelesson.presentation
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import kz.chocofamily.coroutinelesson.R
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
