--- app/src/main/java/kz/chocofamily/coroutinelesson/presentation/extensions/CryptoCurrencyUIExtensions.kt
@@ -1,11 +1,0 @@
-package kz.chocofamily.coroutinelesson.presentation.extensions
-
-import kz.chocofamily.coroutinelesson.data.entities.CryptoCurrencyModel
-
-/**
- * Created by Arslan Tsoy <t.me/arslantsoy> on 2020-04-10
- */
-
-internal fun CryptoCurrencyModel.getUIMarketCap(): String {
-    return ""
-}
