+++ app/src/main/java/kz/example/coroutinelesson/presentation/MainActivity.kt
@@ -1,0 +1,13 @@
+package kz.example.coroutinelesson.presentation
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import kz.example.coroutinelesson.R
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
